[js] rename highlighAnnotations -> highlights
