+++ src/js/directives/highlights.js
