+++ src/js/directives/index.js
@@ -3,7 +3,7 @@ module.exports = function (app) {
   require('./confirmed-click.js')(app);
   require('./create-inline-annotation.js')(app);
   require('./gravatar.js')(app);
-  require('./highlightAnnotations')(app);
+  require('./highlights')(app);
   require('./inline-annotation.js')(app);
   require('./kramjax.js')(app);
   require('./onSelect.js')(app);
